added 128-bit omega1 and omega2 implementation

+++ src/sieve/lpf_mu.cu
@@ -98,7 +98,7 @@ __device__ void atomicMult(int32_t * addr, int32_t val)
 
 __device__ void atomicMult(int64_t * addr, int64_t val)
 {
-  int old = *addr, assumed;
+  int64_t old = *addr, assumed;
 
   do{
     assumed = old;
